--- src/components/SelectLang/SelectLang.module.scss
@@ -1,132 +1,0 @@
-@import '@styles/colors';
-@import '@styles/mixins';
-
-// .lang {
-// 	display: flex;
-// 	flex-direction: row;
-// 	align-items: center;
-// 	margin-left: 50px;
-// 	color: $neutral-text-dark-color;
-// }
-
-// .selectLang {
-// 	appearance: none;
-// 	text-transform: uppercase;
-// 	font-weight: 500;
-// 	margin-left: 10px;
-// 	border: none;
-// 	cursor: pointer;
-// 	@include font-source-sans(15px, $neutral-text-dark-color, 400, 18px, 'Fredoka');
-
-// 	&:hover,
-// 	&:focus {
-// 		outline: none;
-// 	}
-// }
-
-/* reset */
-button {
-	all: unset;
-}
-
-.trigger {
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
-	appearance: none;
-	border: none;
-
-	margin-left: 10px;
-	gap: 10px;
-
-	@include font-source-sans(15px, $neutral-text-dark-color, 400, 18px, 'Fredoka');
-	text-transform: uppercase;
-
-	transition: opacity 0.2s ease-in-out;
-
-	&:hover {
-		opacity: 0.7;
-		cursor: pointer;
-	}
-}
-
-.SelectTrigger:hover {
-	background-color: $primary-background-color;
-}
-.SelectTrigger:focus {
-	box-shadow: 0 0 0 2px black;
-}
-.SelectTrigger[data-placeholder] {
-	color: $primary-button-color;
-}
-
-.SelectIcon {
-	color: $primary-button-dark-color;
-}
-
-.content {
-	overflow: hidden;
-	background-color: white;
-	border-radius: 6px;
-	box-shadow:
-		0px 10px 38px -10px rgba(22, 23, 24, 0.35),
-		0px 10px 20px -15px rgba(22, 23, 24, 0.2);
-}
-
-.viewport {
-	padding: 5px;
-}
-
-.SelectItem {
-	font-size: 13px;
-	line-height: 1;
-	color: $primary-button-dark-color;
-	border-radius: 3px;
-	display: flex;
-	align-items: center;
-	height: 25px;
-	padding: 0 35px 0 25px;
-	position: relative;
-	user-select: none;
-}
-.SelectItem[data-disabled] {
-	color: $neutral-text-light-color;
-	pointer-events: none;
-}
-.SelectItem[data-highlighted] {
-	outline: none;
-	background-color: $neutral-background-second-color;
-	color: $neutral-text-dark-color;
-}
-
-.SelectLabel {
-	padding: 0 25px;
-	font-size: 12px;
-	line-height: 25px;
-	color: $neutral-text-dark-color;
-}
-
-.SelectSeparator {
-	height: 1px;
-	background-color: $neutral-line-color;
-	margin: 5px;
-}
-
-.SelectItemIndicator {
-	position: absolute;
-	left: 0;
-	width: 25px;
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
-}
-
-.SelectScrollButton {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	height: 25px;
-	background-color: white;
-	color: $primary-button-dark-color;
-	cursor: default;
-}
