--- src/components/SelectLang/index.ts
@@ -1,1 +1,0 @@
-export { SelectLang } from './SelectLang';
