+++ src/layouts/LayoutWhithNavbar/LayoutWhithNavbar.tsx
@@ -1,5 +1,5 @@
 import Navbar from '@components/Navbar';
-import Header from '@components/Header';
+import Header from '@/modules/Header';
 import Footer from '@components/Footer';
 import styles from './LayoutWhithNavbar.module.scss';
 import { Outlet } from 'react-router';
